Resolve unmounting instance issue
