+++ config/eslint.config.js
@@ -5,9 +5,10 @@ module.exports = {
   },
   extends: [
     "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
     "plugin:prettier/recommended",
-    "plugin:react/recommended",
-    "plugin:@typescript-eslint/recommended"
+    "plugin:react-hooks/recommended",
+    "plugin:react/recommended"
   ],
   parser: "@typescript-eslint/parser",
   parserOptions: {
@@ -17,20 +18,20 @@ module.exports = {
     ecmaVersion: 12,
     sourceType: "module"
   },
-  plugins: ["prettier", "react", "@typescript-eslint", "simple-import-sort"],
+  plugins: ["@typescript-eslint", "prettier", "react", "simple-import-sort"],
   overrides: [
     {
       files: [".js", ".jsx", ".ts", ".tsx"]
     }
   ],
   rules: {
-    indent: ["error", 2],
+    "@typescript-eslint/no-explicit-any": "off",
     "linebreak-style": ["error", "unix"],
     "simple-import-sort/exports": "error",
     "simple-import-sort/imports": "error",
+    indent: ["error", 2],
     quotes: ["error", "double"],
-    semi: ["error", "always"],
-    "@typescript-eslint/no-explicit-any": "off"
+    semi: ["error", "always"]
   },
   settings: {
     react: {
