+++ package.json
@@ -20,7 +20,8 @@
     "lint:fix": "npm run lint -- --fix",
     "release": "npm run build && gh-pages -d dist/demo && npm publish",
     "start": "webpack serve --config config/webpack.config.js --mode development",
-    "test": "jest --config config/jest.config.js --coverage"
+    "test": "jest --config config/jest.config.js --coverage",
+    "test:watch": "npm run test -- --watch"
   },
   "keywords": [
     "react",
@@ -57,54 +58,55 @@
     "url": "https://github.com/jamesrweb/react-p5-wrapper/issues"
   },
   "dependencies": {
-    "@types/p5": "^1.3.3",
+    "@types/p5": "^1.4.0",
     "microdiff": "^1.3.0",
-    "p5": "^1.4.0",
-    "rooks": "^5.9.0"
+    "p5": "^1.4.1",
+    "rooks": "^5.10.0"
   },
   "peerDependencies": {
     "react": ">=16.8.0",
     "react-dom": ">=16.8.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.16.8",
-    "@babel/core": "^7.16.10",
-    "@babel/preset-env": "^7.16.10",
+    "@babel/cli": "^7.17.0",
+    "@babel/core": "^7.17.2",
+    "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
-    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.2",
     "@types/deep-equal": "^1.0.1",
     "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.38",
+    "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0",
     "babel-loader": "^8.2.3",
     "canvas": "^2.9.0",
-    "css-loader": "^6.5.1",
-    "eslint": "^8.7.0",
+    "css-loader": "^6.6.0",
+    "eslint": "^8.9.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "gh-pages": "^3.2.3",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.4.7",
+    "jest": "^27.5.1",
     "prettier": "^2.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "rimraf": "^3.0.2",
-    "rollup": "^2.64.0",
+    "rollup": "^2.67.2",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.31.1",
+    "rollup-plugin-typescript2": "^0.31.2",
     "style-loader": "^3.3.1",
     "ts-jest": "^27.1.3",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.5.4",
-    "webpack": "^5.66.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.7.3"
+    "typescript": "^4.5.5",
+    "webpack": "^5.68.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   },
   "prettier": {
     "trailingComma": "none"
